adapt to api change - SeekTime

+++ src/client.cpp
@@ -680,7 +680,7 @@ void DemuxReset(void) {}
 void DemuxFlush(void) {}
 PVR_ERROR GetRecordingEdl(const PVR_RECORDING&, PVR_EDL_ENTRY[], int*) { return PVR_ERROR_NOT_IMPLEMENTED; };
 unsigned int GetChannelSwitchDelay(void) { return 0; }
-bool SeekTime(int,bool,double*) { return false; }
+bool SeekTime(double,bool,double*) { return false; }
 void SetSpeed(int) {};
 bool IsTimeshifting(void) { return false; }
 time_t GetPlayingTime() { return 0; }
